Move 'sign in with google' logic to GoogleFlow class
