+++ CMakeLists.txt
@@ -117,6 +117,7 @@ add_library(app STATIC
   src/error_page.cc
   src/gondarsite.cc
   src/gondarwizard.cc
+  src/googleflow.cc
   src/image_select_page.cc
   src/log.cc
   src/meepo.cc
