+++ src/chromeover_login_page.cc
@@ -31,22 +31,6 @@
 #include "oauth_server.h"
 #include "util.h"
 
-// use possibleCharacters to generate a random string
-// based on:
-// https://stackoverflow.com/questions/18862963/qt-c-random-string-generation
-static QString get_random_string() {
-  const QString possibleCharacters(
-      "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789");
-  auto len = 20;
-  QString ret;
-  for (auto i = 0; i < len; i++) {
-    auto index = gondar::getRandomNum(0, possibleCharacters.length() - 1);
-    QChar cur_char = possibleCharacters.at(index);
-    ret.append(cur_char);
-  }
-  return ret;
-}
-
 ChromeoverLoginPage::ChromeoverLoginPage(QWidget* parent) : WizardPage(parent) {
   setTitle("Login");
   setSubTitle("Enter credentials for my.neverware.com.");
@@ -66,11 +50,6 @@ ChromeoverLoginPage::ChromeoverLoginPage(QWidget* parent) : WizardPage(parent) {
   forgotLabel.setTextFormat(Qt::RichText);
   forgotLabel.setTextInteractionFlags(Qt::TextBrowserInteraction);
   forgotLabel.setOpenExternalLinks(true);
-  client_id = gondar::getGoogleSignInId();
-  client_secret = gondar::getGoogleSignInSecret();
-  redirect_uri = QString("http://127.0.0.1:%1").arg(localServer.getPort());
-  state = get_random_string();
-  code_verifier = get_random_string();
   googleLabel.setText("<a href='#'>Sign in with Google</a>");
   googleLabel.setTextFormat(Qt::RichText);
   googleLabel.setTextInteractionFlags(Qt::TextBrowserInteraction);
@@ -83,88 +62,19 @@ ChromeoverLoginPage::ChromeoverLoginPage(QWidget* parent) : WizardPage(parent) {
   meanWordsLabel.setVisible(false);
   // if our build does not have a client id and client secret configured,
   // do not display the 'sign in with google' text
-  if (client_id.isEmpty() || client_secret.isEmpty()) {
-    googleLabel.setVisible(false);
-  }
+  googleLabel.setVisible(googleFlow.shouldShowSignInWithGoogle());
   setLayout(&layout);
   // don't allow progressing to next page yet
   finished = false;
   // don't allow another launch of server interaction while another is running
   started = false;
-  // connect our various event listeners
-  connect(&googleLabel, &QLabel::linkActivated, this,
-          &ChromeoverLoginPage::handleGoogleSigninPart1);
-  connect(&localServer, &OauthServer::authError, this,
-          &ChromeoverLoginPage::handleGoogleSigninError);
-  connect(&localServer, &OauthServer::callbackReceived, this,
-          &ChromeoverLoginPage::handleGoogleSigninPart2);
-  connect(&manager, &QNetworkAccessManager::finished, this,
-          &ChromeoverLoginPage::handleGoogleSigninPart3);
+  // connect events between our members and ourself
+  connect(&googleLabel, &QLabel::linkActivated, &googleFlow,
+          &GoogleFlow::handleGoogleSigninPart1);
   connect(&meepo_, &gondar::Meepo::finished, this,
           &ChromeoverLoginPage::handleMeepoFinished);
-}
-
-void ChromeoverLoginPage::handleGoogleSigninError(QString error) {
-  LOG_WARNING << "received google sign in error=" << error;
-}
-
-// This fires when the user has clicked on the 'sign in with google' link
-// we generate the URL to open in their browser, send them there, and start
-// the local server to parse the sign-in response
-void ChromeoverLoginPage::handleGoogleSigninPart1() {
-  static bool http_server_started = false;
-  if (http_server_started == false) {
-    http_server_started = true;
-    localServer.start();
-    QString googleText =
-        QString(
-            "https://accounts.google.com/o/oauth2/v2/"
-            "auth?client_id=%1&redirect_uri=%2&response_type=code&scope=email&"
-            "state=%3&code_"
-            "challenge_method=plain&code_challenge=%4")
-            .arg(client_id, redirect_uri, state, code_verifier);
-    // notably, using QDesktopServices instead of using a regular hyperlink
-    // appears to suppress the windows warning about opening a local port
-    QDesktopServices::openUrl(QUrl(googleText));
-  }
-}
-
-// This fires when the local server has received the code required to receive
-// a token.  We send a request to the token url, passing along this code
-// and some other identification information to get our token.
-void ChromeoverLoginPage::handleGoogleSigninPart2(QString state_in,
-                                                  QString code) {
-  // make sure we get back the same state string as was sent in initial request
-  if (state_in == state) {
-    LOG_INFO << "Received 'Sign-in with Google' response to local server";
-    QJsonObject json;
-    json["code"] = code;
-    json["client_id"] = client_id;
-    json["client_secret"] = client_secret;
-    json["redirect_uri"] = redirect_uri;
-    json["grant_type"] = "authorization_code";
-    json["code_verifier"] = code_verifier;
-    QJsonDocument doc(json);
-    QUrl tokenUrl = QUrl("https://www.googleapis.com/oauth2/v4/token");
-    QNetworkRequest request(tokenUrl);
-    request.setHeader(QNetworkRequest::ContentTypeHeader, "application/json");
-    manager.post(request, doc.toJson(QJsonDocument::Compact));
-  } else {
-    // TODO(kendall): explicitly show "that's not right..." text to user?
-    LOG_WARNING << "error: state did not match";
-  }
-}
-
-// This fires when we have received our token. We send the token to Meepo
-// to find the user's sites in a way very similar to user/password flow.
-// TODO(kendall): handle failure case
-void ChromeoverLoginPage::handleGoogleSigninPart3(QNetworkReply* reply) {
-  LOG_INFO << "id token received; contacting meepo...";
-  QString id_token = gondar::jsonFromReply(reply)["id_token"].toString();
-  if (!started) {
-    meepo_.startGoogle(id_token);
-    started = true;
-  }
+  connect(googleFlow.getManager(), &QNetworkAccessManager::finished, this,
+          &ChromeoverLoginPage::handleGoogleSigninFinished);
 }
 
 int ChromeoverLoginPage::nextId() const {
@@ -185,7 +95,7 @@ bool ChromeoverLoginPage::validatePage() {
   auth.setUser(usernameLineEdit.text());
   auth.setPassword(passwordLineEdit.text());
   if (finished) {
-    localServer.stop();
+    googleFlow.stopServer();
     return true;
   }
   if (!started) {
@@ -216,5 +126,17 @@ void ChromeoverLoginPage::handleMeepoFinished() {
   } else {
     started = false;
     meanWordsLabel.setVisible(true);
+  }
+}
+
+// This fires when we have received our token. We send the token to Meepo
+// to find the user's sites in a way very similar to user/password flow.
+// TODO(kendall): handle failure case
+void ChromeoverLoginPage::handleGoogleSigninFinished(QNetworkReply* reply) {
+  LOG_INFO << "id token received; contacting meepo...";
+  QString id_token = gondar::jsonFromReply(reply)["id_token"].toString();
+  if (!started) {
+    meepo_.startGoogle(id_token);
+    started = true;
   }
 }
